Renamespace source services
