+++ src/Controller/SourceController.php
@@ -7,9 +7,9 @@ namespace App\Controller;
 use App\Entity\GitSource;
 use App\Repository\SourceRepository;
 use App\Request\GitSourceRequest;
-use App\Services\SourceFactory;
-use App\Services\SourceMutator;
-use App\Services\SourceStore;
+use App\Services\Source\SourceFactory;
+use App\Services\Source\SourceMutator;
+use App\Services\Source\SourceStore;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
