+++ tests/Functional/Services/Source/SourceFactoryTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Functional\Services;
+namespace App\Tests\Functional\Services\Source;
 
 use App\Entity\FileSource;
 use App\Entity\GitSource;
@@ -10,8 +10,8 @@ use App\Entity\RunSource;
 use App\Entity\SourceInterface;
 use App\Repository\SourceRepository;
 use App\Request\GitSourceRequest;
-use App\Services\SourceFactory;
-use App\Tests\Services\SourceRemover;
+use App\Services\Source\SourceFactory;
+use App\Tests\Services\Source\SourceRemover;
 use SmartAssert\UsersSecurityBundle\Security\User;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Symfony\Component\Security\Core\User\UserInterface;
